AArch64: Make graphics classes 64-bit compatible
Changes in this patch include

[x] Long is used to store native pointers as they can
    be 64-bit.

[x] Some minor changes have been done to conform with
    standard JNI practice (e.g. use of jint instead of int
    in JNI function prototypes)

[x] TODO added to AssetAtlasService.java

+++ graphics/java/android/graphics/ComposePathEffect.java
@@ -27,6 +27,6 @@ public class ComposePathEffect extends PathEffect {
                                        innerpe.native_instance);
     }
     
-    private static native int nativeCreate(int outerpe, int innerpe);
+    private static native long nativeCreate(long outerpe, long innerpe);
 }
